Cattle Rounder Western Game
In this western game you are a cowgirl/cowboy. The goal is to Round up all the cattle before they get away. The cattle will start zig zagging part way through the game making it more difficult. The game ends when you miss a single cattle. Your moving abilities are left(arrow) and right(arrow). To use the lasso you press 'enter'. In the top left corner of the screen, is a score board. Your score will increase when you lasso a cow.